+++ test/unit/rabbitmq_class/rabbitmqpub_dropqueue.py
@@ -1,12 +1,12 @@
 #!/usr/bin/python
 # Classification (U)
 
-"""Program:  RabbitMQPub_bindqueue.py
+"""Program:  rabbitmqpub_dropqueue.py
 
-    Description:  Unit test of RabbitMQPub.bind_queue in rabbitmq_class.py.
+    Description:  Unit test of rabbitmqpub.drop_queue in rabbitmq_class.py.
 
     Usage:
-        test/unit/rabbitmq_class/RabbitMQPub_bindqueue.py
+        test/unit/rabbitmq_class/rabbitmqpub_dropqueue.py
 
     Arguments:
 
@@ -34,15 +34,15 @@ import version
 __version__ = version.__version__
 
 
-class BindQueue(object):
+class DropQueue(object):
 
-    """Class:  BindQueue
+    """Class:  DropQueue
 
     Description:  Class stub holder for pika class.
 
     Methods:
         __init__ -> Class initialization.
-        exchange_declare -> Stub holder for exchange_declare function.
+        basic_publish -> Stub holder for basic_publish function.
 
     """
 
@@ -57,25 +57,25 @@ class BindQueue(object):
         """
 
         self.queue = None
-        self.exchange = None
-        self.routing_key = None
+        self.if_unused = None
+        self.if_empty = None
 
-    def queue_bind(self, queue, exchange, routing_key):
+    def queue_delete(self, queue, if_unused, if_empty):
 
-        """Function:  queue_bind
+        """Function:  queue_delete
 
-        Description:  Stub holder for queue_bind function.
+        Description:  Stub holder for queue_delete function.
 
         Arguments:
             queue -> Arg stub holder.
-            exchange -> Arg stub holder.
-            routing_key -> Arg stub holder.
+            if_unused -> Arg stub holder.
+            if_empty -> Arg stub holder.
 
         """
 
         self.queue = queue
-        self.exchange = exchange
-        self.routing_key = routing_key
+        self.if_unused = if_unused
+        self.if_empty = if_empty
 
         return True
 
@@ -88,7 +88,7 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
-        test_bind_queue -> Test bind_queue method.
+        test_drop_queue -> Test drop_queue method.
 
     """
 
@@ -110,13 +110,14 @@ class UnitTest(unittest.TestCase):
         self.queue_name = "Queue_Name"
         self.routing_key = "Route_Key"
         self.auto_delete = True
+        self.body = "Message_Body"
 
     @mock.patch("rabbitmq_class.pika")
-    def test_bind_queue(self, mock_pika):
+    def test_drop_queue(self, mock_pika):
 
-        """Function:  test_bind_queue
+        """Function:  test_drop_queue
 
-        Description:  Test bind_queue method.
+        Description:  Test drop_queue method.
 
         Arguments:
 
@@ -124,10 +125,11 @@ class UnitTest(unittest.TestCase):
 
         mock_pika.PlainCredentials.return_value = "PlainCredentials"
         mock_pika.ConnectionParameters.return_value = "ConnectionParameters"
+        mock_pika.BasicProperties.return_value = True
         rmq = rabbitmq_class.RabbitMQPub(self.name, "xxxxx")
-        rmq.channel = BindQueue()
+        rmq.channel = DropQueue()
 
-        self.assertFalse(rmq.bind_queue())
+        self.assertFalse(rmq.drop_queue())
 
 
 if __name__ == "__main__":